+++ dask/highlevelgraph.py
@@ -220,14 +220,6 @@ class Layer(collections.abc.Mapping):
             v.update(annotations.get(k, {}))
         annotations.update(expanded)
 
-    @property
-    def layer_materialize_module(self):
-        return type(self).__module__
-
-    @property
-    def layer_materialize_class(self):
-        return type(self).__name__
-
     def clone(
         self,
         keys: set,
@@ -943,8 +935,8 @@ class HighLevelGraph(Mapping):
         for layer in (self.layers[name] for name in self._toposort_layers()):
             layers.append(
                 {
-                    "__module__": layer.layer_materialize_module,
-                    "__name__": layer.layer_materialize_class,
+                    "__module__": layer.__module__,
+                    "__name__": type(layer).__name__,
                     "state": layer.__dask_distributed_pack__(
                         self.get_all_external_keys(),
                         self.key_dependencies,
